round robin and rematch finished

+++ app/Http/Controllers/V1/PoolsController.php
@@ -51,8 +51,7 @@ class PoolsController extends Controller
         foreach($timeTable as $timeTableCategory) {
             $timeTableCategories [] = $timeTableCategory->category_id;
         }
- 
-        
+
         $registrations = $compatition->registrations;
         $reg_single = $registrations->where('team_or_single', 1)->whereIn('category_id', $timeTableCategories)->countBy('category_id');
         $reg_teams = $registrations->where('team_or_single', 0)->whereIn('category_id', $timeTableCategories)->countBy('category_id');
@@ -73,8 +72,10 @@ class PoolsController extends Controller
         foreach($reg_teams as $key=>$count){
             $nn_team_cat[] = $registrations->where('category_id', $key)->groupBy('team_id')->values();
         }
-  
-        
+
+        if($compatition->registrations->where('position', '!=', NULL)->count() > 0) {
+            return $this->error('', 'Takmicenje je vec pocelo', 422);
+        }
         if($pools->count() > 0 ) {
             Pool::destroy($pools);
             
@@ -86,9 +87,8 @@ class PoolsController extends Controller
         if($timeTable->count() == 0) {
             return $this->error('', 'Potrebno je prvo da se odredi Time Table', 422);
         }
-        if($compatition->registrations->where('position', '!=', NULL)->count() > 0) {
-            return $this->error('', 'Takmicenje je vec pocelo', 422);
-        }
+
+
         /** Here we start rebuilding */
         //return $nn_team_cat;
         
@@ -237,7 +237,7 @@ class PoolsController extends Controller
             }
         }
         //return response()->json(!in_array($pool->pool_type, ['FM', 'SF', 'RFM', 'RSF']));
-        if(!in_array($pool->pool_type, ['FM', 'SF', 'RFM', 'RSF'])) {
+        if(!in_array($pool->pool_type, ['FM', 'SF', 'REFM'])) {
             $request->looserId != 'null' ? $looserRegistration->update(['position' => null]) : null;
             $winnerRegistration->update(['position' => null]);
         }
@@ -277,16 +277,33 @@ class PoolsController extends Controller
                 $timeTable->update(['status'=> 2, 'finish_time' => Date("H:i:s", strtotime(now()))]);
             }
             if($compatition->rematch == 1) {
-                if($category->repesaz == 0) {
+                if($category->repesaz == 0 || ($category->repesaz == 1 && $category->kata_or_kumite == 1)) {
                     $timeTable->update(['status'=> 2, 'finish_time' => Date("H:i:s", strtotime(now()))]);
-                } else {
+                } 
+                if($category->repesaz == 1 && $category->kata_or_kumite == 0) {
+                    $request->winnerId != 'null' ? $pool->update(['winner_id' => $winnerId]) : $pool->update(['winner_id' => null]);
+                    $request->looserId != 'null' ? $pool->update(['looser_id' => $looserId]) : $pool->update(['looser_id' => null]);
                     $repesazData = $this->rematchBuilding($pool);
+                    $oldRepData = Pool::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id)->whereIn('pool_type', ['RE', 'REFM']);
+                    if($oldRepData->get()->count() > 0){
+                        $oldRepData->delete();
+                    }
                     Pool::insert($repesazData);
                 }
             }
         }
         
         //Repasaz
+        if($pool->pool_type == 'RE') {
+            $request->nextMatchId != 'null' ? $nextMetch->update(['registration_one' => $winnerId]) : null;
+        }
+        if($pool->pool_type == 'REFM'){
+            $winnerId != 'null' ? $winnerRegistration->update(['position' => '1']) : null;
+            $timeTable->update(['status'=> 2, 'finish_time' => Date("H:i:s", strtotime(now()))]);
+        }
+        // if($pool->pool_type == 'FM' && $compatition->rematch == 1) {
+        //     if($category->repesaz == 1)
+        // }
         // if($pool->pool_type == 'FM' && $compatition->rematch == 1 && $category->repesaz == 1 ) {
         //     $getPool = Pool::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id);
         //     $getLoosers = $getPool->where('winner_id', $winnerId)->where('looser_id', '!=', null)->where('looser_id', '!=', $request->looserId)->orderBy('pool', 'asc')->get();
@@ -329,99 +346,99 @@ class PoolsController extends Controller
             
             
         // }
-        if($pool->pool_type == 'RR') {
-            $request->winnerId != 'null' ? $pool->update(['winner_id' => $winnerId]) : $pool->update(['winner_id' => null]);
-            $request->looserId != 'null' ? $pool->update(['looser_id' => $looserId]) : $pool->update(['looser_id' => null]);
-            $poolsRR = Pool::where('compatition_id', $pool->compatition_id)->where('category_id',$pool->category_id)->where('pool_type', 'RR')->get();
-            $poolsRRFM = Pool::where('compatition_id', $pool->compatition_id)->where('category_id',$pool->category_id)->where('pool_type', 'RRFM')->first();
-            $groupOne = [1,2,3];
-            $groupTwo = [4,5,6];
-            $totalRRPools = $poolsRR->count();
+        // if($pool->pool_type == 'RR') {
+        //     $request->winnerId != 'null' ? $pool->update(['winner_id' => $winnerId]) : $pool->update(['winner_id' => null]);
+        //     $request->looserId != 'null' ? $pool->update(['looser_id' => $looserId]) : $pool->update(['looser_id' => null]);
+        //     $poolsRR = Pool::where('compatition_id', $pool->compatition_id)->where('category_id',$pool->category_id)->where('pool_type', 'RR')->get();
+        //     $poolsRRFM = Pool::where('compatition_id', $pool->compatition_id)->where('category_id',$pool->category_id)->where('pool_type', 'RRFM')->first();
+        //     $groupOne = [1,2,3];
+        //     $groupTwo = [4,5,6];
+        //     $totalRRPools = $poolsRR->count();
 
-            $totalFinishedGroupOne = $poolsRR->whereIn('group', $groupOne)->where('winner_id', '!=', 'null')->count();
-            $totalFinishedGroupTwo = $poolsRR->whereIn('group', $groupTwo)->where('winner_id', '!=', 'null')->count();
-            $registeredRR = Registration::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id)->get();
-            if($totalRRPools == 3 && $totalFinishedGroupOne == 3) {
-                //$registeredRR = Registration::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id)->get();
+        //     $totalFinishedGroupOne = $poolsRR->whereIn('group', $groupOne)->where('winner_id', '!=', 'null')->count();
+        //     $totalFinishedGroupTwo = $poolsRR->whereIn('group', $groupTwo)->where('winner_id', '!=', 'null')->count();
+        //     $registeredRR = Registration::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id)->get();
+        //     if($totalRRPools == 3 && $totalFinishedGroupOne == 3) {
+        //         //$registeredRR = Registration::where('compatition_id', $pool->compatition_id)->where('category_id', $pool->category_id)->get();
                 
 
-            }
-            if($totalRRPools == 6 ) {
-                if($pool->group <= 3 && $totalFinishedGroupOne == 3) {
-                    $firstPlaceRRGroupOne = null;
-                    $secondPlaceRRGroupOne = null;
-                    $thirdPlaceRRGroupOne = null;
+        //     }
+        //     if($totalRRPools == 6 ) {
+        //         if($pool->group <= 3 && $totalFinishedGroupOne == 3) {
+        //             $firstPlaceRRGroupOne = null;
+        //             $secondPlaceRRGroupOne = null;
+        //             $thirdPlaceRRGroupOne = null;
 
-                    foreach($registeredRR as $regCompa) {
-                        $winingCount = $poolsRR->whereIn('group', $groupOne)->where('winner_id', $regCompa->id)->count();
-                        $loosingCount = $poolsRR->whereIn('group', $groupOne)->where('looser_id', $regCompa->id)->count();
+        //             foreach($registeredRR as $regCompa) {
+        //                 $winingCount = $poolsRR->whereIn('group', $groupOne)->where('winner_id', $regCompa->id)->count();
+        //                 $loosingCount = $poolsRR->whereIn('group', $groupOne)->where('looser_id', $regCompa->id)->count();
 
-                        $loosingCount == 2 ? $thirdPlaceRRGroupOne = $regCompa->id : null;
-                        switch($winingCount){
-                            case 2:
-                                $firstPlaceRRGroupOne = $regCompa->id;
-                                break;
-                            case 1:
-                                $secondPlaceRRGroupOne = $regCompa->id;
-                                break;
-                        }
-                    }
-                    $registrationsPositionCount = Registration::whereIn('id', [$firstPlaceRRGroupOne, $secondPlaceRRGroupOne, $thirdPlaceRRGroupOne])->get();
-                    if($firstPlaceRRGroupOne != null) {
-                        $poolsRRFM->update(['registration_one' => $firstPlaceRRGroupOne]);
-                    }
-                    if($secondPlaceRRGroupOne != null) {
-                        $registrationUpdate = Registration::where('id', $secondPlaceRRGroupOne)->first();
-                        $registrationUpdate->update(['position' => '1']);
-                    }
-                    return $registrationUpdate;
-                }
-                if($pool->group > 3 && $totalFinishedGroupTwo == 3) {
-                    $firstPlaceRRGroupTwo = null;
-                    $secondPlaceRRGroupTwo = null;
-                    $thirdPlaceRRGroupTwo = null;
+        //                 $loosingCount == 2 ? $thirdPlaceRRGroupOne = $regCompa->id : null;
+        //                 switch($winingCount){
+        //                     case 2:
+        //                         $firstPlaceRRGroupOne = $regCompa->id;
+        //                         break;
+        //                     case 1:
+        //                         $secondPlaceRRGroupOne = $regCompa->id;
+        //                         break;
+        //                 }
+        //             }
+        //             $registrationsPositionCount = Registration::whereIn('id', [$firstPlaceRRGroupOne, $secondPlaceRRGroupOne, $thirdPlaceRRGroupOne])->get();
+        //             if($firstPlaceRRGroupOne != null) {
+        //                 $poolsRRFM->update(['registration_one' => $firstPlaceRRGroupOne]);
+        //             }
+        //             if($secondPlaceRRGroupOne != null) {
+        //                 $registrationUpdate = Registration::where('id', $secondPlaceRRGroupOne)->first();
+        //                 $registrationUpdate->update(['position' => '1']);
+        //             }
+        //             return $registrationUpdate;
+        //         }
+        //         if($pool->group > 3 && $totalFinishedGroupTwo == 3) {
+        //             $firstPlaceRRGroupTwo = null;
+        //             $secondPlaceRRGroupTwo = null;
+        //             $thirdPlaceRRGroupTwo = null;
 
 
-                    foreach($registeredRR as $regCompa) {
-                        $winingCount = $poolsRR->whereIn('group', $groupTwo)->where('winner_id', $regCompa->id)->count();
-                        $loosingCount = $poolsRR->whereIn('group', $groupTwo)->where('looser_id', $regCompa->id)->count();
-                        $loosingCount == 2 ? $thirdPlaceRRGroupTwo = $regCompa->id : null;
+        //             foreach($registeredRR as $regCompa) {
+        //                 $winingCount = $poolsRR->whereIn('group', $groupTwo)->where('winner_id', $regCompa->id)->count();
+        //                 $loosingCount = $poolsRR->whereIn('group', $groupTwo)->where('looser_id', $regCompa->id)->count();
+        //                 $loosingCount == 2 ? $thirdPlaceRRGroupTwo = $regCompa->id : null;
 
-                        switch($winingCount){
-                            case 2:
-                                $firstPlaceRRGroupTwo = $regCompa->id;
-                                break;
-                            case 1:
-                                $secondPlaceRRGroupTwo = $regCompa->id;
-                                break;
-                        }
-                    }
-                    $registrationsPositionCount = Registration::whereIn('id', [$firstPlaceRRGroupTwo, $secondPlaceRRGroupTwo, $thirdPlaceRRGroupTwo])->where('position', '!=', 'null')->get();
-                    return $registrationsPositionCount->count();
-                    if($firstPlaceRRGroupTwo != null) {
-                        $poolsRRFM->update(['registration_two' => $firstPlaceRRGroupTwo]);
-                    }
-                    if($secondPlaceRRGroupTwo != null) {
-                        $registrationUpdate = Registration::where('id', $secondPlaceRRGroupTwo)->first();
-                        $registrationUpdate->update(['position' => '1']);
-                    }
-                    return $registrationUpdate;
-                }
+        //                 switch($winingCount){
+        //                     case 2:
+        //                         $firstPlaceRRGroupTwo = $regCompa->id;
+        //                         break;
+        //                     case 1:
+        //                         $secondPlaceRRGroupTwo = $regCompa->id;
+        //                         break;
+        //                 }
+        //             }
+        //             $registrationsPositionCount = Registration::whereIn('id', [$firstPlaceRRGroupTwo, $secondPlaceRRGroupTwo, $thirdPlaceRRGroupTwo])->where('position', '!=', 'null')->get();
+        //             return $registrationsPositionCount->count();
+        //             if($firstPlaceRRGroupTwo != null) {
+        //                 $poolsRRFM->update(['registration_two' => $firstPlaceRRGroupTwo]);
+        //             }
+        //             if($secondPlaceRRGroupTwo != null) {
+        //                 $registrationUpdate = Registration::where('id', $secondPlaceRRGroupTwo)->first();
+        //                 $registrationUpdate->update(['position' => '1']);
+        //             }
+        //             return $registrationUpdate;
+        //         }
                 
-                return "$registeredRR 111";
+        //         return "$registeredRR 111";
 
-            }
-            return 1;
-            if($totalRRPools == 5) {
-                //
-            }
+        //     }
+        //     return 1;
+        //     if($totalRRPools == 5) {
+        //         //
+        //     }
 
-        }
-        if($pool->pool_type == 'REFM') {
-            $request->looserId != 'null' ? $looserResult =  1 : $looserResult = null;
-            Registration::where('id', $looserId)->first()->update(['position' => $looserResult]);
-            Registration::where('id', $winnerId)->first()->update(['position' => null]);
-        }
+        // }
+        // if($pool->pool_type == 'REFM') {
+        //     $request->looserId != 'null' ? $looserResult =  1 : $looserResult = null;
+        //     Registration::where('id', $looserId)->first()->update(['position' => $looserResult]);
+        //     Registration::where('id', $winnerId)->first()->update(['position' => null]);
+        // }
         // if($pool->pool_type == 'RFM') {
         //     $request->looserId != 'null' ? $looserResult =  1 : $looserResult = null;
         //     $request->winnerId != 'null' ? $winnerResult =  2 : $winnerResult = null;
@@ -434,7 +451,6 @@ class PoolsController extends Controller
             $request->looserId != 'null' ? $pool->update(['looser_id' => $looserId]) : $pool->update(['looser_id' => null]);
         }
         //update results
-        //return 22;
         $this->calculateResults($pool->compatition_id ,array_unique($clubsArray), 'results');
     
         return new PoolResource($pool);
